+++ lib/mod.js
@@ -1,4 +1,4 @@
-import Shim, { domShimSymbol, defaultGlobals } from './shim.js';
+import setupShim, { domShimSymbol, defaultGlobals } from './shim.js';
 
 const thisModuleURL = new URL(import.meta.url);
 
@@ -9,7 +9,7 @@ if (thisModuleURL.searchParams.has('props')) {
   shimedGlobals = defaultGlobals;
 }
 
-const { shim, unshim } = Shim(shimedGlobals);
+let { shim, unshim } = setupShim(shimedGlobals);
 
 if (thisModuleURL.searchParams.has('global')) {
   shim();
