+++ lib/shim.js
@@ -1,4 +1,5 @@
 import { parseHTML } from 'https://unpkg.com/linkedom@0.11.0/worker.js';
+import { apply as applyPatches } from './patch.js';
 
 export const domShimSymbol = Symbol.for('dom-shim.defaultView');
 export const defaultGlobals = Object.freeze([
@@ -9,78 +10,14 @@ export const defaultGlobals = Object.freeze([
 
 const dontShimIfExists = new Set(['window']);
 
-export default function (shimedGlobals = Object.freeze([])) {
+export default function(shimedGlobals = Object.freeze([])) {
   let parsed = parseHTML(`
     <html lang="en">
     <head><title>Site</title></head>
     <body></body>
     </html>
   `);
-
-  // Patch Text
-  let aTextNode = parsed.document.createTextNode('');
-  let ConstructibleText = aTextNode.constructor.bind(null, parsed.document);
-
-  // Patch Node
-  const CharacterData = Object.getPrototypeOf(ConstructibleText);
-  const Node = Object.getPrototypeOf(CharacterData);
-  const EventTarget = Object.getPrototypeOf(Node);
-
-  const ntSymbol = Symbol('dom-shim.nodeType');
-  class PatchedNode extends EventTarget {
-    constructor() {
-      super();
-
-      Object.defineProperty(this, 'nodeType', {
-        get() {
-          return this[ntSymbol];
-        },
-        set(val) {
-          this[ntSymbol] = val;
-        },
-      });
-    }
-  }
-
-  Object.setPrototypeOf(Node, PatchedNode);
-  Object.setPrototypeOf(Node.prototype, PatchedNode.prototype);
-
-  // Patch requestAnimationFrame
-  let lastTime = 0;
-  function requestAnimationFrame(callback, element) {
-    var currTime = new Date().getTime();
-    var timeToCall = Math.max(0, 16 - (currTime - lastTime));
-    var id = setTimeout(function () {
-      callback(currTime + timeToCall);
-    }, timeToCall);
-    lastTime = currTime + timeToCall;
-    return id;
-  }
-
-  const dfDescriptor = Object.getOwnPropertyDescriptor(
-    parsed.Document.prototype,
-    'defaultView'
-  );
-  Object.defineProperty(parsed.Document.prototype, 'defaultView', {
-    get() {
-      let window = dfDescriptor.get.call(this);
-
-      return new Proxy(window, {
-        get(window, name) {
-          switch (name) {
-            case 'Text':
-              return ConstructibleText;
-            case 'requestAnimationFrame':
-              return requestAnimationFrame;
-          }
-          return window[name];
-        },
-      });
-    },
-  });
-
-  // Patch HTMLElement
-  delete parsed.HTMLElement.observedAttributes;
+  applyPatches(parsed);
 
   const window = parsed.document.defaultView;
 
@@ -113,6 +50,6 @@ export default function (shimedGlobals = Object.freeze([])) {
 
   return {
     shim,
-    unshim,
+    unshim
   };
 }
